checkbox - resolved requisicao post erro 403 - resolved tabela clientes - resolved botao excel - resolved

+++ frontend/src/services/Endpoints.js
@@ -3,28 +3,32 @@ import axios from "axios";
 const httpClient = axios.create({
   baseURL: "http://localhost:8080",
   headers: {
-    "Content-type": "application/json"
-  }
-})
+    "Content-type": "application/json",
+  },
+});
 
 export const getAll = () => {
-  return httpClient.get("/home/clientes", { headers: { Authorization: "Bearer " + localStorage.getItem("jwtToken") } });
+  return httpClient.get("/home/clientes", {
+    headers: { Authorization: "Bearer " + localStorage.getItem("jwtToken") },
+  });
 };
 
-export const get = nome => {
-  return httpClient.get(`/home/clientes/${nome}`,{});
+export const get = (nome) => {
+  return httpClient.get(`/home/clientes/${nome}`, {
+    headers: { Authorization: "Bearer " + localStorage.getItem("jwtToken") },
+  });
 };
 
-export const create = data => {
+export const create = (data) => {
   return httpClient.post("/client", data);
 };
 
 export const update = (id, data) => {
-  return httpClient.put(`/home/clientes/${id}`, data);
+  return httpClient.put(`/home/clientes/${id}`, data, {headers: { Authorization: "Bearer " + localStorage.getItem("jwtToken") }});
 };
 
-export const remove = id => {
-  return httpClient.delete(`/home/clientes/${id}`);
+export const remove = (id) => {
+  return httpClient.delete(`/home/clientes/${id}`, {headers: { Authorization: "Bearer " + localStorage.getItem("jwtToken") }});
 };
 
 export const exportExcel = () => {
